requirements.txt
mysql-connector-python
gspread
pandas